smart class
